+++ omnigibson/prims/joint_prim.py
@@ -885,7 +885,7 @@ class JointPrim(BasePrim):
     def _serialize(self, state):
         # We serialize by iterating over the keys and adding them to a list that's concatenated at the end
         # This is a deterministic mapping because we assume the state is an OrderedDict
-        return np.concatenate(list(state.values()))
+        return np.concatenate(list(state.values())).astype(float)
 
     def _deserialize(self, state):
         # We deserialize deterministically by knowing the order of values -- pos, vel, effort
